Move ServiceBase to services package
